PI-3853 Fix v3 tier counts JSON mapping (#759)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppstier/jpa/repository/TierSummaryRepository.kt
@@ -25,5 +25,5 @@ interface TierSummaryRepository : JpaRepository<TierSummaryEntity, String> {
         order by ts.tier
         """
     )
-    fun getTierV3Counts(): List<Pair<String, Int>>
+    fun getTierV3Counts(): List<Pair<String, Long>>
 }
